Docker compose for web and database DONE
The error was caused by not executing the "mvnw package -DskipTests" command after modifying the scripts in vscode, so the init method remained commented.

After regenerating the jar file with the PostConstruct method enabled and the database option in update mode, the app works perfectly.

Also, the commit includes a yml file to execute from external users with no code, only having docker installed. The images are uploaded into a public repository on Docker Hub so that anyone can download them executing the docker-compose.yml script and run the app.

There are some things that can be upgraded though:
- The web app does not wait until the database is ready
- The needed port for the database is the 3306, which collides with host's MySQL port.